backend/requirements.txt
# FastAPI and web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.20
aiofiles>=24.0.0

# Data validation and settings
pydantic
pydantic-settings

# Database
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
alembic>=1.14.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# ML and Embeddings
sentence-transformers
# torch installed separately via --index-url
transformers

# Vector Database
qdrant-client>=1.7.0

# LLM Integration
groq>=0.4.1

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
pandas>=2.1.0
openpyxl>=3.1.2

# Caching (optional)
redis>=5.0.1

# Utilities
httpx>=0.26.0
tenacity>=8.2.3
